HACK: serial: 8250_dw: skip lpss pci UART D3 when used with no_console_suspend
For Platforms with lpss pci UARTs, the parent device should
be prevented from going into D3 for the no_console_suspend
flag to work as expected.

This is tested with Intel SPT platforms with lpss pci UART devices.

Signed-off-by: Archana Patni <archana.patni@intel.com>
Signed-off-by: Subramony Sesha <subramony.sesha@intel.com>

Interim. This solution is not ideal. We should not be
concerned of the parent in this driver, and we really should
not do any PCI operations here.

Ideally the issue can be fixed in the parent driver
(drivers/mfd/intel-lpss-pci.c).

Signed-off-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>