REVERTME:SKL:Topology: Add logic to create SDW aggregation blob.
Blob needs to be sent to copier for SDW aggregation. Normally
blob is derived from NHLT, but since in SDW blob, stream number
is dynamic, blob needs to be created in driver.  Blob creation
logic for the SDW is implemented in this function. Actually
this should go in skl-messages.c as per review. Need to move
this to skl-messages.c in next patch set.

Change-Id: Ifc58a2343498190f736295754be2c14e3c6d5bea
Signed-off-by: Hardik Shah <hardik.t.shah@intel.com>
Reviewed-on: https://localhost/gerrit/9416