SKL:PCM: Derive the SDW master controller number from be_id.
SDW Master number or SSP instance number can be derived from
VBUS ID. But for aggregation one copier is linked to two
SDW master controllers, so its not feasible to get controller
number from vbus_id. So fixing the range of BE_IDs for number
of controllers to find for which controller PCM ops got called.

Change-Id: I6cb995404e0488e660f35f943e9bfc3fdc518e07
Signed-off-by: Hardik Shah <hardik.t.shah@intel.com>
Reviewed-on: https://localhost/gerrit/9417