ASoC: Intel: Skylake: Fix incorrect parsing of pipe tokens
To avoid parsing of pipe related tokens multiple times for a case where
the pipe has more than one module, a logic was added to parse these
tokens only once. But, the existing logic would parse these
tokens only if there are more than one module in the pipe. So, for a
pipe with single module, the logic is insufficient.

So, fix it by updating the flag accordingly.

Change-Id: Ie183f14eaf98f21cf87691e0e681b77342706b37
Signed-off-by: Shreyas NC <shreyas.nc@intel.com>
Signed-off-by: Diwakar, Praveen <praveen.diwakar@intel.com>
Reviewed-on: https://localhost/gerrit/15665
Reviewed-by: audio_build <audio_build@intel.com>
Reviewed-by: Prodduvaka, Leoni <leoni.prodduvaka@intel.com>
Reviewed-by: S, Pavan K <pavan.k.s@intel.com>
Reviewed-by: R, Dharageswari <dharageswari.r@intel.com>
Reviewed-by: Singh, Guneshwor O <guneshwor.o.singh@intel.com>
Reviewed-by: Kale, Sanyog R <sanyog.r.kale@intel.com>
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Tested-by: Sm, Bhadur A <bhadur.a.sm@intel.com>