drm/i915/gvt: add module parameter enable_pvmmio
This bool type module parameter is used to enable/disable the pvmmio
feature for MMIO emulation in GVT. This parameter is default true.
Its permission type is 0400 which means user could only change its
value through the cmdline, this is to prevent the dynamic modification
during runtime which would break the pvmmio internal logic.

Notice: this patch is required to be applied to guest kernel.

Change-Id: I570f1fe02101e518595c02fce67601b692871aa9
Signed-off-by: Pei Zhang <pei.zhang@intel.com>
Acknowledged-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>
Reviewed-on: https://localhost/613366
Reviewed-by: He, Min <min.he@intel.com>
Reviewed-by: Jiang, Fei <fei.jiang@intel.com>
Reviewed-by: Dong, Eddie <eddie.dong@intel.com>
Tested-by: Dong, Eddie <eddie.dong@intel.com>