ASoC: Intel: Skylake: Add support for GAIN module
Gain module is represented by three entities such as Ramp Duration,
Ramp Type and Volume controlled by three kcontrols. These kcontrols
are added with the corresponding get and put handlers.

A Large IPC message is sent to the DSP when any of these controls
are set and if widget is in powered on state else these values
are cached in driver and sent during init of gain module.

Multichannel support for volume control is implemented by
user defined kcontrol info method to set number
of channels to what we are getting from mconfig out channels.

Change-Id: I2473d636fb673f7a73118d7b5018f07127def69f
Signed-off-by: Giribabu Gogineni <giribabux.gogineni@intel.com>
Signed-off-by: Dharageswari.R <dharageswari.r@intel.com>
Reviewed-on: https://localhost/gerrit/16929
Reviewed-by: Sinha, Mohit <mohit.sinha@intel.com>
Reviewed-by: Shaik, Kareem M <kareem.m.shaik@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Tested-by: Sm, Bhadur A <bhadur.a.sm@intel.com>