drm/i915/gvt: Change DomU to support 3 HDMI displays.
Request DomU supports 3 virtual HDMI displays.

This requires eDP-1, DP-1 and DP-2 to be disabled on DomU command line.
DomU's ias.conf must also be updated to support HDMI and not DP displays.
If a panel is not present on port A, DDI B and DDI C will be made active
(as in Dom0).

Reviewed by: Min He <min.he@intel.com>
Signed-off-by: Michael Byrne <michaelx.byrne@intel.com>

Change-Id: Ib944de7bf9be5c19db28d35edd77a9f89148f27a
Reviewed-on: https://localhost/613357
Reviewed-by: He, Min <min.he@intel.com>
Reviewed-by: Jiang, Fei <fei.jiang@intel.com>
Reviewed-by: Dong, Eddie <eddie.dong@intel.com>
Tested-by: Dong, Eddie <eddie.dong@intel.com>